add device already exists functionality
